requirements.txt
Appium-Python-Client==3.1.1
pytest==7.4.0
allure-pytest==2.13.2
pytest-order==1.1.0
